Refactoring QueryPage
Ticket: #309
