+++ src/app/Main/Body/QueryPage/QueryPageContent/DatabaseStatusIndicator/QueryPage-DatabaseStatusIndicator.component.tsx
@@ -2,9 +2,9 @@
 import { css, jsx } from "@emotion/core";
 import { useTranslation } from "react-i18next";
 import { useEffect, useState } from 'react';
-import { callApiService, ApiResponse } from '../../../../Core/callApi.service';
-import { DatabaseStatusDTO } from "../../../../Shared/Model/Api_Database.model";
-import { DATABASE_STATUS_URL } from "../../../../Shared/URLs";
+import { callApiService, ApiResponse } from '../../../../../Core/callApi.service';
+import { DatabaseStatusDTO } from "../../../../../Shared/Model/Api_Database.model";
+import { DATABASE_STATUS_URL } from "../../../../../Shared/URLs";
 
 const statusStyle = css`
     text-align: center;
