+++ src/app/Main/Body/QueryPage/QueryPageContent/IntroText/QueryPage-IntroText.component.tsx
