+++ src/app/Main/Body/QueryPage/QueryPageContent/Results/Charts/ApexChart.model.ts
