+++ src/app/Main/Body/QueryPage/QueryPageContent/Results/TableResults/QueryPageContent-TableResultsLayout.component.tsx
@@ -3,10 +3,10 @@ import { css, jsx } from "@emotion/core";
 import { useState } from "react";
 import { useTranslation } from "react-i18next";
 import _ from "lodash";
-import { AccordionComponent } from "../../../../../Shared/Accordion.component";
+import { AccordionComponent } from "../../../../../../Shared/Accordion.component";
 import { ResultsTableOptionsComponent } from "./TableResults-Options.component";
-import { ExplanationTextComponent } from "../../../../../Shared/ExplanationText.component";
-import { DataInterface } from "../../../../../Shared/Context/DataContext";
+import { ExplanationTextComponent } from "../../../../../../Shared/ExplanationText.component";
+import { DataInterface } from "../../../../../../Shared/Context/DataContext";
 import { TableResultsTableComponent } from "./TableResults-Table.component";
 import { SumOptions } from "./TableResults.model";
