+++ src/app/Main/Body/QueryPage/QueryPageContent/Results/TableResults/TableContent-RowWithColSums.component.tsx
