+++ src/app/Main/Body/QueryPage/QueryPageContent/Results/TableResults/TableResults.model.ts
