+++ src/app/Shared/Model/Client_Isolate.model.ts
@@ -63,7 +63,6 @@ export const MainFilterList: DbKey[] = [
     "productionType",
     "matrix",
     "federalState",
-    "resistance",
 ];
 
 export const subFiltersList: string[] = ["serovar", "species", "serotyp"];
